+++ app/src/main/java/com/example/dacn/Lich_su_lam_bai/Fragment/GeoFragment.java
@@ -1,0 +1,21 @@
+package com.example.dacn.Lich_su_lam_bai.Fragment;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.dacn.R;
+
+public class GeoFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_geo, container, false);
+    }
+}
