--- app/src/main/java/com/example/dacn/Lich_su_lam_bai/Lich_su_lam_bai_model.java
@@ -1,22 +1,0 @@
-package com.example.dacn.Lich_su_lam_bai;
-
-public class Lich_su_lam_bai_model {
-    String ten_mon_hoc;
-    String ten_de_thi;
-
-    public Lich_su_lam_bai_model(String ten_mon_hoc, String ten_de_thi) {
-        this.ten_mon_hoc = ten_mon_hoc;
-        this.ten_de_thi = ten_de_thi;
-
-    }
-
-    public String getTen_mon_hoc() {
-        return ten_mon_hoc;
-    }
-
-    public String getTen_de_thi() {
-        return ten_de_thi;
-    }
-
-
-}
