+++ app/src/main/res/layout/activity_subject_history.xml
@@ -1,23 +1,58 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:orientation="vertical"
+    android:background="@color/hong">
 
-    >
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/toolbar_history"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:background="@color/timxanh"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent">
+
+        <ImageView
+            android:id="@+id/img_back_veapp"
+            android:layout_width="30dp"
+            android:layout_height="30dp"
+            android:src="@drawable/ic_back"
+            app:tint="@color/trang"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            android:layout_marginStart="15dp"/>
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:fontFamily="@font/roboto_bold"
+            android:text="Lịch sử làm bài"
+            android:textColor="@color/trang"
+            android:textSize="26sp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
     <com.google.android.material.tabs.TabLayout
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
-        android:id="@+id/tabLayout">
+        android:id="@+id/tabLayout"
+        android:background="@color/timnhat"
+        android:padding="3dp"
+        app:tabIndicatorColor="@color/den80"
+        app:tabTextColor="@color/den80">
 
         <com.google.android.material.tabs.TabItem
             android:layout_height="wrap_content"
             android:layout_width="match_parent"
-            android:text="Sử"
-            android:textColor = "@color/timxanh"
-            />
+            android:text="Sử"/>
 
         <com.google.android.material.tabs.TabItem
             android:layout_height="wrap_content"
@@ -43,8 +78,4 @@
         android:id="@+id/view_page1"
         android:layout_below="@+id/tabLayout"/>
 
-
-
-
-
-</RelativeLayout>
+</LinearLayout>
