+++ app/src/main/res/layout/fragment_eng.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Lich_su_lam_bai.Subject_history.EngFragment">
+    tools:context=".Lich_su_lam_bai.Fragment.EngFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
